+++ __tests__/__system__/cli/issue/cli.issue.system.test.ts
@@ -8,12 +8,11 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-import {runCliScript} from "../../../__src__/TestUtils";
-import {ITestEnvironment} from "../../../__src__/environment/doc/response/ITestEnvironment";
-import {TestEnvironment} from "../../../__src__/environment/TestEnvironment";
+import { ITestEnvironment, TestEnvironment, runCliScript } from "@zowe/cli-test-utils";
+import { ITestPropertiesSchema } from "../../../__src__/environment/doc/ITestPropertiesSchema";
 
 // Test environment will be populated in the "beforeAll"
-let TEST_ENVIRONMENT: ITestEnvironment;
+let TEST_ENVIRONMENT: ITestEnvironment<ITestPropertiesSchema>;
 
 describe("zowe-cli-sample issue command", () => {
 
@@ -27,8 +26,8 @@ describe("zowe-cli-sample issue command", () => {
     afterAll(async () => {
         await TestEnvironment.cleanUp(TEST_ENVIRONMENT);
     });
-    it("should display the help", async () => {
-        const response = await runCliScript(__dirname + "/__scripts__/issue_help.sh", TEST_ENVIRONMENT);
+    it("should display the help", () => {
+        const response = runCliScript(__dirname + "/__scripts__/issue_help.sh", TEST_ENVIRONMENT);
 
         expect(response.stderr.toString()).toBe("");
         expect(response.status).toBe(0);