+++ src/cli/list/profile-args/ProfileArgs.definition.ts
@@ -9,7 +9,7 @@
  */
 
 import { ICommandDefinition } from "@zowe/imperative";
-import { ZosmfSession } from "@zowe/zosmf-for-zowe-sdk";
+import { ProfileConstants } from "../../../api/ProfileConstants";
 /**
  * Command one [object] definition. This definition is of imperative type "command" and therefore must have a
  * command handler (which performs the "work" for this command).
@@ -25,24 +25,17 @@ import { ZosmfSession } from "@zowe/zosmf-for-zowe-sdk";
  * "handler" is the file path to the handler (does the work)
  * "options" an array of options
  */
-export const DirectoryContentsDefinition: ICommandDefinition = {
-    name: "directory-contents",
-    aliases: ["dc"],
-    summary: "Lists directory contents",
+export const ProfileArgsDefinition: ICommandDefinition = {
+    name: "profile-args",
+    aliases: ["pa"],
+    summary: "Lists profile args",
     description: "[objects] in Zowe CLI are the entities on which [actions] are performed. [objects] are always nouns. " +
         "For example, for command \"zowe zos-files delete data-set\", the [object] is \"data-set\".\n\nFor this command, " +
         "we list contents of the specified directory.",
     type: "command",
-    handler: __dirname + "/DirectoryContents.handler",
-    positionals: [
-        {
-            name: "directory",
-            description: "The directory/path to list the contents. Blank/Omitted will list the current directory",
-            type: "string"
-        }
-    ],
+    handler: __dirname + "/ProfileArgs.handler",
     profile: {
-        optional: ["zosmf"],
+        optional: ["sample"],
     },
-    options: [...ZosmfSession.ZOSMF_CONNECTION_OPTIONS]
+    options: [...ProfileConstants.SAMPLE_CONNECTION_OPTIONS]
 };